Move user entity, service and controller to User Module
