+++ src/app.module.ts
@@ -5,11 +5,13 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { getOrmConfig } from './utils/getOrmConfig';
 import { ContentModule } from './content/content.module';
 import { SocialModule } from './social/social.module';
+import { UserModule } from './user/user.module';
 @Module({
   imports: [
     TypeOrmModule.forRoot(getOrmConfig(process.env.NODE_ENV)),
     ContentModule,
     SocialModule,
+    UserModule,
     // TypeOrmModule.forFeature([Sculpture, SculptureMaker]),
   ],
   controllers: [AppController],
