+++ src/content/content.module.ts
@@ -6,22 +6,14 @@ import { SculptureMakerService } from './service/sculpture-maker.service';
 import { SculptureImageService } from './service/image.service';
 import { PictureUploader } from './service/picture-uploader.service';
 import { SculptureImageController } from './controller/image.controller';
-import { UserService } from './service/user.service';
-import { UserController } from './controller/user.controller';
 
 @Module({
-  controllers: [
-    SculptureController,
-    MakerController,
-    SculptureImageController,
-    UserController,
-  ],
+  controllers: [SculptureController, MakerController, SculptureImageController],
   providers: [
     SculptureService,
     SculptureMakerService,
     SculptureImageService,
     PictureUploader,
-    UserService,
   ],
 })
 export class ContentModule {}
